enforce windows 1252 encoding for .csv files
After changing the gitattributes file, i opened every single csv file in the localisation folder and reopened it with the windows 1252 encoding, then made a minor change like add or remove an empty line or some unneeded ; to save that and make github recognise a change and also set the end of line encoding to LF. Github would then also show all the places that are affected by the encoding change, most of them are fine, but a few were not. I fixed the ones i could, but some usually french or spanish loc still contains § instead of letters like é, ê, î or whatever. Aside from maybe the LegacyOfRome and ReapersDue nothing should cause problems though and in those cases it is likely also only the non english localisation that is affected and now contains a few hundred ».